# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Ptium API
  version: 0.1.0
  summary: AI presentation generation, administration, and MCP API
  description: |
    Ptium exposes owner-scoped presentation generation, profile and API-key
    management, administrator operations, and a stateless MCP endpoint.

    JSON REST responses use an envelope containing either `data` or `error`
    plus `requestId`. List responses additionally contain `meta`. The `/mcp`
    endpoint uses raw JSON-RPC 2.0 messages and does not use the REST envelope.

    JSON request bodies are limited to 2 MiB, reject unknown top-level fields,
    and must contain exactly one complete JSON value. UUID path parameters are
    validated before their operation handler runs.

    OIDC access tokens and Ptium API keys are both accepted as Bearer
    credentials. Ptium API keys may alternatively be sent in `X-API-Key`.
    Development authentication is disabled by default and uses
    `X-Ptium-Dev-Secret` unless the server is configured with another header.

    The `x-ptium-api-key-scopes` extension on an operation names the scope a
    Ptium API key must grant. Interactive OIDC and development sessions are
    not restricted by API-key scopes. MCP API-key calls always require
    `mcp:use`; each presentation tool additionally requires its documented
    `presentations:read` or `presentations:write` scope.
servers:
  - url: /
    description: Same origin as the Ptium web application
...
security:
  - BearerAuth: []
  - APIKeyHeader: []
  - DevSecret: []
paths:
...
  /api/v1/profile:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    get:
      operationId: getProfile
      summary: Get the current user's personalization profile
      tags: [Profile]
      x-ptium-api-key-scopes: [profile:read]
      responses:
        '200':
          description: Personalization profile.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      operationId: replaceProfile
      summary: Replace the current user's personalization profile
      description: Omitted fields are stored as empty strings and preferences default to an empty object.
      tags: [Profile]
      x-ptium-api-key-scopes: [profile:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileWrite'
      responses:
        '200':
          description: Updated personalization profile.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateProfile
      summary: Update the current user's personalization profile
      description: |
        This endpoint currently has the same replacement behavior as PUT;
        omitted scalar fields are cleared and omitted preferences become `{}`.
      tags: [Profile]
      x-ptium-api-key-scopes: [profile:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileWrite'
      responses:
        '200':
          description: Updated personalization profile.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/presentations/{id}:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
      - $ref: '#/components/parameters/PresentationID'
    get:
      operationId: getPresentation
      summary: Get a presentation and its slides
      tags: [Presentations]
      x-ptium-api-key-scopes: [presentations:read]
      responses:
        '200':
          description: Presentation owned by the current user.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      operationId: replacePresentation
      summary: Update presentation metadata and optionally replace slides
      description: |
        Metadata fields are merged with the current presentation. When `slides`
        is present, the entire slide sequence is atomically replaced; array
        order determines the saved `position`.
      tags: [Presentations]
      x-ptium-api-key-scopes: [presentations:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresentationUpdate'
      responses:
        '200':
          description: Updated presentation including its current slides.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updatePresentation
      summary: Partially update presentation metadata or replace slides
      description: When `slides` is present, the entire sequence is replaced; array order is authoritative.
      tags: [Presentations]
      x-ptium-api-key-scopes: [presentations:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresentationUpdate'
      responses:
        '200':
          description: Updated presentation including its current slides.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deletePresentation
      summary: Move a presentation to the recycle bin
      tags: [Presentations]
      x-ptium-api-key-scopes: [presentations:write]
      responses:
        '204':
          description: Presentation moved to the recycle bin. The response has no body.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/presentations/{id}/source:
    parameters:
      - $ref: '#/components/parameters/PresentationID'
      - $ref: '#/components/parameters/XRequestID'
    get:
      operationId: getPresentationSource
      summary: Read the deck as source
      description: |
        Returns the deck written in Ptium's slide language, together with the
        layout ids and component names this template accepts. The source is the
        deck: compiling it produces exactly the stored slides.

        When the slides have been edited on the canvas since the source was
        written, the text is regenerated from them, so the two never disagree.

        The language is documented in `docs/deck-source.md`.
      tags: [Presentations]
      responses:
        '200':
          description: The deck as text.
          content:
            application/json:
              schema:
                type: object
                required: [data, requestId]
                properties:
                  data:
                    type: object
                    properties:
                      source:
                        type: string
                      slideCount:
                        type: integer
                      language:
                        type: string
                      blockKinds:
                        type: array
                        items:
                          type: string
                      layouts:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            name:
                              type: string
                            role:
                              type: string
                              enum: [title, section, content, twoContent, comparison, quote, picture, table, chart, closing, blank]
                  requestId:
                    $ref: '#/components/schemas/RequestID'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      operationId: applyPresentationSource
      summary: Compile deck source into slides
      description: |
        Compiles source against the deck's template and replaces its slides.

        Source that asks for something the template cannot do is adjusted rather
        than rejected — a missing layout, a component with no room, text with
        nowhere to go — and every adjustment is reported in `warnings`.

        With `dryRun` the deck is left untouched and only the result is reported,
        which is what an editor uses to check before applying.
      tags: [Presentations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              additionalProperties: false
              properties:
                source:
                  type: string
                  description: The deck in Ptium's slide language.
                dryRun:
                  type: boolean
                  default: false
                version:
                  type: integer
                  format: int64
                  minimum: 1
                  description: Optional expected version when applying. Ignored for a dry run.
      responses:
        '200':
          description: The compiled deck, applied unless this was a dry run.
          content:
            application/json:
              schema:
                type: object
                required: [data, requestId]
                properties:
                  data:
                    type: object
                    properties:
                      applied:
                        type: boolean
                      warnings:
                        type: array
                        items:
                          type: string
                      presentation:
                        $ref: '#/components/schemas/Presentation'
                      slides:
                        type: array
                        description: Present on a dry run only.
                        items:
                          $ref: '#/components/schemas/Slide'
                  requestId:
                    $ref: '#/components/schemas/RequestID'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          description: The source produced no slides, more than the limit, or is not text.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/presentations/{id}/inspect:
    parameters:
      - $ref: '#/components/parameters/PresentationID'
      - $ref: '#/components/parameters/XRequestID'
    get:
      operationId: inspectPresentation
      summary: Measure a deck as it will be drawn
      description: |
        Answers the question someone would otherwise answer by exporting the file
        and looking at every slide: does anything overflow, collide, fall off the
        edge, or become unreadable.

        The same measurements run inside every source compile, whose response
        carries `findings` beside its warnings.
      tags: [Presentations]
      responses:
        '200':
          description: What the deck looks like once drawn.
          content:
            application/json:
              schema:
                type: object
                required: [data, requestId]
                properties:
                  data:
                    type: object
                    properties:
                      slides:
                        type: integer
                      clean:
                        type: boolean
                        description: True when no defect was found; advisories do not affect it.
                      defects:
                        type: integer
                      advisories:
                        type: integer
                      findings:
                        type: array
                        items:
                          $ref: '#/components/schemas/DeckFinding'
                  requestId:
                    $ref: '#/components/schemas/RequestID'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The deck has no slides to measure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OIDC JWT or ptium_* API key
      description: |
        Send `Authorization: Bearer <credential>`. The credential may be an
        OIDC access-token JWT validated through issuer discovery and rotating
        JWKS, or an opaque Ptium API key beginning with `ptium_`.
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Alternative header for a Ptium `ptium_*` API key. Do not send it
        together with Authorization. The key must grant the scope named by the
        operation's `x-ptium-api-key-scopes` extension.
    DevSecret:
      type: apiKey
      in: header
      name: X-Ptium-Dev-Secret
      description: |
        Development-only shared secret. It is disabled by default, normally
        restricted to loopback clients, and may be renamed by server
        configuration. A Bearer value beginning `dev:` may accompany it to
        select an allowed development identity.
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Caller-generated correlation ID. Values longer than 128 characters are replaced.
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    PresentationID:
      name: id
      in: path
      required: true
      description: Presentation UUID.
      schema:
        type: string
        format: uuid
...
  headers:
    XRequestID:
      description: Correlation ID used by logs, errors, and administrator incidents.
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    BadRequest:
      description: Malformed, oversized, trailing, or multi-value JSON; unknown fields; an invalid UUID path ID; or invalid transport input.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            invalidJSON:
              value:
                error:
                  code: invalid_json
                  message: Request body is not valid for this operation
                  details:
                    reason: 'json: unknown field "unexpected"'
                requestId: 8dcb8033f7d1f8027f2df88d
            invalidUUID:
              value:
                error:
                  code: invalid_id
                  message: Path id must be a UUID
                requestId: 8dcb8033f7d1f8027f2df88d
    Unauthorized:
      description: A credential is absent, invalid, expired, or revoked.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        WWW-Authenticate:
          description: Bearer authentication challenge.
          schema:
            type: string
            example: Bearer realm="ptium"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: authentication_required
              message: Unauthorized
            requestId: 8dcb8033f7d1f8027f2df88d
    Forbidden:
      description: The account is disabled or the API key lacks the operation's required scope.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            insufficientScope:
              value:
                error:
                  code: insufficient_scope
                  message: The API key does not grant this operation
                  details:
                    required: presentations:write
                requestId: 8dcb8033f7d1f8027f2df88d
            disabled:
              value:
                error:
                  code: account_disabled
                  message: This account has been disabled
                requestId: 8dcb8033f7d1f8027f2df88d
...
    NotFound:
      description: The record does not exist or is not owned by the current user.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: not_found
              message: The requested resource was not found
            requestId: 8dcb8033f7d1f8027f2df88d
    Conflict:
      description: Current resource state conflicts with the requested operation.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            emptyPresentation:
              value:
                error:
                  code: presentation_has_no_slides
                  message: Generate or add slides before exporting
                requestId: 8dcb8033f7d1f8027f2df88d
            selfAdministration:
              value:
                error:
                  code: cannot_disable_self
                  message: Administrators cannot disable or demote their own active account
                requestId: 8dcb8033f7d1f8027f2df88d
    UnprocessableEntity:
      description: Syntactically valid input failed operation validation.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            fieldValidation:
              value:
                error:
                  code: validation_error
                  message: requestedSlideCount must be between 1 and 50
                requestId: 8dcb8033f7d1f8027f2df88d
            settingValidation:
              value:
                error:
                  code: validation_error
                  message: API key grace must be a duration between zero and 30 days
                  details:
                    key: security.api_key_grace
                requestId: 8dcb8033f7d1f8027f2df88d
            settingRelationship:
              value:
                error:
                  code: validation_error
                  message: generation default slide count cannot exceed the maximum
                requestId: 8dcb8033f7d1f8027f2df88d
    InternalError:
      description: An unexpected server or database error. The error is captured for administrators.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: internal_error
              message: The server could not complete the request
            requestId: 8dcb8033f7d1f8027f2df88d
...
  schemas:
    JSONValue:
      description: Any valid JSON value.
    RequestID:
      type: string
      minLength: 1
      maxLength: 128
      example: 8dcb8033f7d1f8027f2df88d
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, requestId]
      properties:
        error:
          $ref: '#/components/schemas/APIError'
        requestId:
          $ref: '#/components/schemas/RequestID'
...
    ProfileWrite:
      type: object
      additionalProperties: false
      properties:
        displayName:
          type: string
          maxLength: 120
          default: ''
        company:
          type: string
          maxLength: 200
          default: ''
        jobTitle:
          type: string
          maxLength: 200
          default: ''
        bio:
          type: string
          maxLength: 4000
          default: ''
        preferences:
          $ref: '#/components/schemas/JSONValue'
      example:
        displayName: Mina Kim
        company: Ptium Labs
        jobTitle: Product strategist
        bio: Builds concise executive narratives.
        preferences:
          preferredLanguage: ko
          preferredTheme: aurora
    ProfileEnvelope:
      type: object
      additionalProperties: false
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/Profile'
        requestId:
          $ref: '#/components/schemas/RequestID'
...
    PresentationEnvelope:
      type: object
      additionalProperties: false
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/Presentation'
        requestId:
          $ref: '#/components/schemas/RequestID'
...